user_service/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
# psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.0.1
redis==5.2.1
django-redis==5.4.0
drf-yasg==1.21.8
celery==5.4.0